move class to elastic aggregations module
